pyproject.toml
[tool.poetry]
name = "hunt-art"
version = "0.1.0"
description = ""
authors = ["Alexey Bobrov <bobrov.bobrov-alex@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
gunicorn = "^21.2.0"
psycopg2 = "^2.9.9"
dj-database-url = "^2.1.0"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}
drf-spectacular = "^0.26.2"
drf-spectacular-sidecar = "^2023.4.1"
pillow = "^10.2.0"
drf-nested-routers = "^0.93.5"
django = "^5.0.2"
django-debug-toolbar = "^4.3.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
